rename /configs to /cfg
